# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Station API
  description: |
    REST API for Station - Laravel queue management package.

    Authentication is configurable via `station.api.auth`:
    - `token` — requires `Authorization: Bearer <token>` header
    - `none` — no authentication (development only)
  version: 0.1.0
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  contact:
    name: OJ Baeza
    url: https://github.com/ojbaeza/station
servers:
  - url: '{baseUrl}/api/station'
    description: Station API
    variables:
      baseUrl:
        default: http://localhost
        description: Application base URL
...
paths:
...
  /jobs/{id}/retry:
    post:
      tags: [Jobs]
      summary: Retry a job
      operationId: retryJob
      parameters:
        - $ref: '#/components/parameters/JobId'
      responses:
        '200':
          description: Job queued for retry
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '404':
          $ref: '#/components/responses/NotFound'
  /jobs/{id}/cancel:
    post:
      tags: [Jobs]
      summary: Cancel a job
      operationId: cancelJob
      parameters:
        - $ref: '#/components/parameters/JobId'
      responses:
        '200':
          description: Job cancelled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '404':
          $ref: '#/components/responses/NotFound'
  /jobs/bulk/cancel:
    post:
      tags: [Jobs]
      summary: Cancel multiple jobs
      operationId: bulkCancelJobs
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BulkIdsRequest'
      responses:
        '200':
          description: Bulk cancel result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BulkActionResponse'
        '422':
          $ref: '#/components/responses/ValidationError'
  /jobs/bulk/retry:
    post:
      tags: [Jobs]
      summary: Retry multiple jobs
      operationId: bulkRetryJobs
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BulkIdsRequest'
      responses:
        '200':
          description: Bulk retry result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BulkActionResponse'
        '422':
          $ref: '#/components/responses/ValidationError'
  /jobs/bulk/delete:
    post:
      tags: [Jobs]
      summary: Delete multiple jobs
      operationId: bulkDeleteJobs
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BulkIdsRequest'
      responses:
        '200':
          description: Bulk delete result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BulkActionResponse'
        '422':
          $ref: '#/components/responses/ValidationError'

  # ── Failed Jobs ────────────────────────────────────────────────
...
  /failed/{id}/retry:
    post:
      tags: [Failed Jobs]
      summary: Retry a failed job
      operationId: retryFailedJob
      parameters:
        - $ref: '#/components/parameters/FailedJobId'
      responses:
        '200':
          description: Failed job queued for retry
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '404':
          $ref: '#/components/responses/NotFound'
  /failed/{id}:
    delete:
      tags: [Failed Jobs]
      summary: Delete a failed job
      operationId: deleteFailedJob
      parameters:
        - $ref: '#/components/parameters/FailedJobId'
      responses:
        '200':
          description: Failed job deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /failed/bulk/retry:
    post:
      tags: [Failed Jobs]
      summary: Retry multiple failed jobs
      operationId: bulkRetryFailed
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BulkIdsRequest'
      responses:
        '200':
          description: Bulk retry result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BulkActionResponse'
        '422':
          $ref: '#/components/responses/ValidationError'
  /failed/bulk/delete:
    post:
      tags: [Failed Jobs]
      summary: Delete multiple failed jobs
      operationId: bulkDeleteFailed
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BulkIdsRequest'
      responses:
        '200':
          description: Bulk delete result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BulkActionResponse'
        '422':
          $ref: '#/components/responses/ValidationError'

  # ── Batches ────────────────────────────────────────────────────
...
  /batches/{id}/cancel:
    post:
      tags: [Batches]
      summary: Cancel a batch
      operationId: cancelBatch
      parameters:
        - $ref: '#/components/parameters/BatchId'
      responses:
        '200':
          description: Batch cancelled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '404':
          $ref: '#/components/responses/NotFound'
  /batches/{id}/retry:
    post:
      tags: [Batches]
      summary: Retry failed jobs in a batch
      operationId: retryBatch
      parameters:
        - $ref: '#/components/parameters/BatchId'
      responses:
        '200':
          description: Batch jobs queued for retry
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '404':
          $ref: '#/components/responses/NotFound'
  /batches/bulk/cancel:
    post:
      tags: [Batches]
      summary: Cancel multiple batches
      operationId: bulkCancelBatches
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BulkIdsRequest'
      responses:
        '200':
          description: Bulk cancel result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BulkActionResponse'
        '422':
          $ref: '#/components/responses/ValidationError'
  /batches/bulk/retry:
    post:
      tags: [Batches]
      summary: Retry multiple batches
      operationId: bulkRetryBatches
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BulkIdsRequest'
      responses:
        '200':
          description: Bulk retry result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BulkActionResponse'
        '422':
          $ref: '#/components/responses/ValidationError'
  /batches/bulk/delete:
    post:
      tags: [Batches]
      summary: Delete multiple batches
      operationId: bulkDeleteBatches
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BulkIdsRequest'
      responses:
        '200':
          description: Bulk delete result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BulkActionResponse'
        '422':
          $ref: '#/components/responses/ValidationError'

  # ── Monitoring ─────────────────────────────────────────────────
...
components:
  parameters:
    JobId:
      name: id
      in: path
      required: true
      schema:
        type: string
      description: Job ID
    FailedJobId:
      name: id
      in: path
      required: true
      schema:
        type: string
      description: Failed job ID
    BatchId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: Batch ID
...
  schemas:
...
    BulkIdsRequest:
      type: object
      required: [ids]
      properties:
        ids:
          type: array
          items:
            type: string
          minItems: 1
          maxItems: 100
          description: Array of resource IDs (max 100)
...
    BulkActionResponse:
      type: object
      properties:
        success:
          type: boolean
        processed:
          type: integer
          description: Number of items successfully processed
        failed:
          type: integer
          description: Number of items that failed
        errors:
          type: array
          items:
            type: object
            properties:
              id:
                type: string
              message:
                type: string

    # ── Common ─────────────────────────────────────────────────
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
        count:
          type: integer
          description: Number of items affected (when applicable)
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
    ValidationErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
        errors:
          type: object
          additionalProperties:
            type: array
            items:
              type: string
  responses:
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            message: The requested resource was not found
...
    ValidationError:
      description: Validation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ValidationErrorResponse'
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: |
        Bearer token authentication. Set `STATION_API_TOKEN` in your `.env`
        and configure `station.api.auth` to `token`.

    sanctum:
      type: apiKey
      in: cookie
      name: laravel_session
      description: Laravel Sanctum session authentication
security:
  - bearerAuth: []
  - sanctum: []